# api-contracts/p0/system-service.openapi.yaml
openapi: 3.1.0
info:
  title: Biel Life Camp System Service API
  version: 1.0.0
  description: P0 身份、组织、RBAC、第三方配置和操作审计契约。
servers:
  - url: /
...
paths:
  /api/system/v1/auth/wechat/login:
    post:
      tags: [Authentication]
      operationId: loginWithWechat
      summary: loginWithWechat
      description: 后端始终以 loginCode 向微信换取 OpenID。已有活动绑定时 phoneCode 可省略；没有绑定时必须提供 phoneCode，且手机号只能唯一匹配本地 EHR 在职员工。
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WechatLoginRequest'
      responses:
        '200':
          description: 已绑定员工登录成功；首次登录仅在微信手机号唯一匹配本地 EHR 在职员工时自动绑定并成功。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenPairApiResponse'
        '400':
          $ref: './common.yaml#/components/responses/BadRequest'
        '403':
          $ref: './common.yaml#/components/responses/Forbidden'
        '409':
          $ref: './common.yaml#/components/responses/Conflict'
        '503':
          description: EHR 初始全量同步尚未成功，登录入口失败关闭。
          content:
            application/json:
              schema:
                $ref: './common.yaml#/components/schemas/ApiResponse'
  /api/system/v1/auth/admin/login:
    post:
      tags: [Authentication]
      operationId: loginAdminWithPassword
      summary: 管理后台密码登录
      description: 过渡和应急登录入口，默认关闭。账号不存在与密码错误统一返回 AUTH_INVALID_CREDENTIALS。
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminLoginRequest'
      responses:
        '200':
          description: 登录成功，签发 ADMIN_WEB/password 会话。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenPairApiResponse'
        '400':
          $ref: './common.yaml#/components/responses/BadRequest'
        '401':
          $ref: './common.yaml#/components/responses/Unauthorized'
        '503':
          description: 本地密码登录未启用或登录限流依赖不可用。
          content:
            application/json:
              schema:
                $ref: './common.yaml#/components/schemas/ApiResponse'
...
  /api/system/v1/auth/token/refresh:
    post:
      tags: [Authentication]
      operationId: refreshAccessToken
      summary: refreshAccessToken
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshTokenRequest'
      responses:
        '200':
          description: 刷新成功；旧刷新令牌同时失效。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenPairApiResponse'
        '401':
          $ref: './common.yaml#/components/responses/Unauthorized'
  /api/system/v1/auth/logout:
    post:
      tags: [Authentication]
      operationId: logoutCurrentSession
      summary: logoutCurrentSession
      security:
        - ExternalBearer: []
      parameters:
        - $ref: './common.yaml#/components/parameters/IdempotencyKey'
      responses:
        '200':
          description: 当前会话已撤销；重复调用仍返回统一成功响应。
          content:
            application/json:
              schema:
                $ref: './common.yaml#/components/schemas/EmptyApiResponse'
        '401':
          $ref: './common.yaml#/components/responses/Unauthorized'
  /api/system/v1/auth/logout-all:
    post:
      tags: [Authentication]
      operationId: logoutAllSessions
      summary: logoutAllSessions
      security:
        - ExternalBearer: []
      parameters:
        - $ref: './common.yaml#/components/parameters/IdempotencyKey'
      responses:
        '200':
          description: 当前员工全部会话已撤销。
          content:
            application/json:
              schema:
                $ref: './common.yaml#/components/schemas/EmptyApiResponse'
        '401':
          $ref: './common.yaml#/components/responses/Unauthorized'
  /api/system/v1/me:
    get:
      tags: [Authentication]
      operationId: getCurrentSubject
      summary: getCurrentSubject
      security:
        - ExternalBearer: []
      responses:
        '200':
          description: 当前员工及授权摘要。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CurrentSubjectApiResponse'
        '401':
          $ref: './common.yaml#/components/responses/Unauthorized'
...
  /internal/system/v1/auth/session-context:
    post:
      tags: [Internal]
      operationId: resolveGatewaySessionContext
      summary: Gateway 在每次外部请求上校验可撤销会话、账户状态和授权版本
      security:
        - ServiceMtls: []
          GatewayServiceToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SessionContextRequest'
      responses:
        '200':
          description: 按目标服务裁剪后的当前内部身份上下文。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionContextApiResponse'
        '401':
          $ref: './common.yaml#/components/responses/Unauthorized'
        '403':
          $ref: './common.yaml#/components/responses/Forbidden'
        '409':
          $ref: './common.yaml#/components/responses/Conflict'
        '503':
          description: system-service 授权状态不可用，Gateway 必须失败关闭。
          content:
            application/json:
              schema:
                $ref: './common.yaml#/components/schemas/ApiResponse'
...
components:
  securitySchemes:
    ExternalBearer:
      $ref: './common.yaml#/components/securitySchemes/ExternalBearer'
    InternalIdentity:
      $ref: './common.yaml#/components/securitySchemes/InternalIdentity'
    ServiceMtls:
      $ref: './common.yaml#/components/securitySchemes/ServiceMtls'
    GatewayServiceToken:
      type: apiKey
      in: header
      name: X-Gateway-Service-Token
...
  schemas:
    TokenPairApiResponse:
      allOf:
        - $ref: './common.yaml#/components/schemas/ApiResponse'
        - type: object
          properties:
            code:
              const: '0'
            errorMsg:
              const: ''
            data:
              $ref: '#/components/schemas/TokenPair'
    CurrentSubjectApiResponse:
      allOf:
        - $ref: './common.yaml#/components/schemas/ApiResponse'
        - type: object
          properties:
            code:
              const: '0'
            errorMsg:
              const: ''
            data:
              $ref: '#/components/schemas/CurrentSubject'
...
    SessionContextApiResponse:
      allOf:
        - $ref: './common.yaml#/components/schemas/ApiResponse'
        - type: object
          properties:
            code:
              const: '0'
            errorMsg:
              const: ''
            data:
              $ref: '#/components/schemas/SessionContextResponse'
...
    TokenPair:
      type: object
      additionalProperties: false
      required: [tokenType, accessToken, accessExpiresIn, refreshToken, refreshExpiresAt]
      properties:
        tokenType:
          type: string
          const: Bearer
        accessToken:
          type: string
          writeOnly: true
        accessExpiresIn:
          type: integer
          const: 900
        refreshToken:
          type: string
          writeOnly: true
        refreshExpiresAt:
          $ref: './common.yaml#/components/schemas/UtcDateTime'
    WechatLoginRequest:
      type: object
      additionalProperties: false
      required: [loginCode]
      properties:
        loginCode:
          type: string
          writeOnly: true
          description: wx.login 获取的一次性登录码。
          minLength: 1
          maxLength: 512
        phoneCode:
          type: string
          writeOnly: true
          description: 首次绑定时 getPhoneNumber 获取的一次性手机号授权码；已有 AppID + OpenID 活动绑定时可省略，客户端不得改为提交明文手机号或 OpenID。
          minLength: 16
          maxLength: 512
    AdminLoginRequest:
      type: object
      additionalProperties: false
      required: [employeeNo, password]
      properties:
        employeeNo:
          type: string
          minLength: 1
          maxLength: 64
        password:
          type: string
          writeOnly: true
          minLength: 1
          maxLength: 128
...
    RefreshTokenRequest:
      type: object
      additionalProperties: false
      required: [refreshToken]
      properties:
        refreshToken:
          type: string
          writeOnly: true
          minLength: 32
    CurrentSubject:
      type: object
      additionalProperties: false
      required: [employeeId, employeeNo, displayName, organizationId, roles, permissions, dataScopes, authzVersion]
      properties:
        employeeId:
          $ref: './common.yaml#/components/schemas/DistributedId'
        employeeNo:
          type: string
        displayName:
          type: string
        organizationId:
          $ref: './common.yaml#/components/schemas/DistributedId'
        roles:
          type: array
          items:
            type: string
        permissions:
          type: array
          items:
            type: string
        dataScopes:
          type: array
          items:
            $ref: './common.yaml#/components/schemas/DataScope'
        authzVersion:
          type: integer
          minimum: 1
...
    SessionContextRequest:
      type: object
      additionalProperties: false
      required: [employeeId, sessionId, authzVersion, targetService]
      properties:
        employeeId:
          $ref: './common.yaml#/components/schemas/DistributedId'
        sessionId:
          type: string
          format: uuid
        authzVersion:
          type: integer
          format: int64
          minimum: 1
        targetService:
          type: string
    SessionContextResponse:
      type: object
      additionalProperties: false
      required: [employeeId, organizationId, sessionId, clientType, authzVersion, roleCodes, permissions, dataScopes, amr]
      properties:
        employeeId:
          $ref: './common.yaml#/components/schemas/DistributedId'
        organizationId:
          $ref: './common.yaml#/components/schemas/DistributedId'
        sessionId:
          type: string
          format: uuid
        clientType:
          type: string
          enum: [MINI_PROGRAM, ADMIN_WEB]
        authzVersion:
          type: integer
          format: int64
          minimum: 1
        roleCodes:
          type: array
          items: { type: string }
        permissions:
          type: array
          items: { type: string }
        dataScopes:
          type: array
          items:
            $ref: './common.yaml#/components/schemas/DataScope'
        amr:
          type: array
          items: { type: string }
...